requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
langchain-huggingface
faiss-cpu
google-generativeai
python-dotenv
beautifulsoup4
fastapi
uvicorn
requests